fasdfa

+++ Assets/_Scripts/Player/MultiPowerUp.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class MultiPowerUp : PowerUp {
 
 	[SerializeField] private float _addMulti = 0.2f;
-	[SerializeField] private ulong _addScore = 10;
+	[SerializeField] private int _addScore = 10;
 
 	public override void OnTriggerEnter2D (Collider2D other)
 	{
